app/requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
langchain-core>=0.1.0
langchain-community>=0.0.0
langchain-ollama>=0.1.0
ollama>=0.1.0
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0
pandas>=2.0.0
